test(instrumentation): convert Spring boot test apps to simple Java apps
We do not actually need to start a whole framework with an HTTP server
to check one environment variable and exit.

Also:
- disable all auto-instrumentation by the Java agent for JVM
  instrumentation image tests
- remove Node.js 18 Bullseye base image for Node.js tests

All of this should help to speed up the instrumentation image tests.
