--- images/instrumentation/test/jvm/.m2/.gitignore
@@ -1,1 +1,0 @@
-repository/*
