+++ images/instrumentation/test/jvm/test-cases/.gitignore
@@ -1,0 +1,1 @@
+*.class
