+++ images/instrumentation/test/jvm/test-cases/existing-unmodified/MANIFEST.MF
@@ -1,0 +1,2 @@
+Manifest-Version: 1.0
+Main-Class: Main
