--- images/instrumentation/test/jvm/test-cases/existing-unmodified/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-spring.application.name=existing-envvar-demo
