+++ images/instrumentation/test/jvm/test-cases/otel-resource-attributes-already-set/.env
