--- images/instrumentation/test/jvm/test-cases/otel-resource-attributes-already-set/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-spring.application.name=otel-resource-attributes-already-set-demo
