+++ images/instrumentation/test/jvm/test-cases/undefined-for-non-existing/MANIFEST.MF
@@ -1,0 +1,2 @@
+Manifest-Version: 1.0
+Main-Class: Main
