--- images/instrumentation/test/jvm/test-cases/undefined-for-non-existing/src/test/java/com/dash0/injector/existing_envvar_demo/TestApplicationTests.java
@@ -1,13 +1,0 @@
-package com.dash0.injector.existing_envvar_demo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TestApplicationTests {
-
-	@Test
-	void contextLoadSucceeds() {
-	}
-
-}
